requirements.txt
numpy
psutil
wmi
pyyaml
pywin32